Add NCIT terms for mzQC (#109)
* Add NCIT terms for mzQC

As per #105, remove the full NCIT import and just re-declare the 4 NCIT terms that are used in mzQC.

I've used abridged terms from NCIT now that contain the relevant information, without some of the unnecessary `property_value` specifications. I'm not sure though whether the `is_a` specifications might be problematic now, because they refer to NCIT parent terms that are not included in our CV? It shouldn't be our goal to redefine the NCIT tree up to those terms, so if this is indeed problematic, what would be the solution? Have the terms consist of just `id`-`name`-`def`?

* Update OBO version